refactor(mcp): finish splitting tool families
