+++ crates/aether-mcp/src/tools/tests/mod.rs
@@ -1,0 +1,12 @@
+mod bytes;
+mod capture;
+mod components;
+mod describe;
+mod engine;
+mod envelope;
+mod ids;
+mod logs_cost;
+mod mail;
+mod render;
+mod reply;
+mod state;
